src/probabilistic_model.py
import pyro
import pyro.distributions as dist
import torch
from typing import List, Tuple
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def mathematical_concept_model(input_data: torch.Tensor) -> Tuple[List[torch.Tensor], torch.Tensor]:
    num_levels = pyro.param("num_levels", torch.tensor(3.0), constraint=dist.constraints.positive)
    
    concepts = []
    for level in range(int(num_levels)):
        level_concepts = generate_level_concepts(level, concepts)
        concepts.append(level_concepts)
    
    observations = generate_observations(concepts, input_data)
    
    return concepts, observations

def generate_level_concepts(level: int, lower_concepts: List[torch.Tensor]) -> torch.Tensor:
    num_concepts = pyro.sample(f"num_concepts_{level}", dist.Poisson(5.0))
    concepts = []
    for i in range(int(num_concepts)):
        if level == 0:
            # Base level concepts are simple transformations
            concept_type = pyro.sample(f"concept_type_{level}_{i}", dist.Categorical(torch.ones(4)))
            if concept_type == 0:  # Identity
                concept = lambda x: x
            elif concept_type == 1:  # Square
                concept = lambda x: x**2
            elif concept_type == 2:  # Exponential
                concept = lambda x: torch.exp(x)
            else:  # Logarithm
                concept = lambda x: torch.log(torch.abs(x) + 1e-8)
        else:
            # Higher level concepts combine lower level concepts
            lower_level_concepts = lower_concepts[-1]
            num_components = pyro.sample(f"num_components_{level}_{i}", dist.Poisson(2.0))
            components = [lower_level_concepts[j] for j in torch.randint(len(lower_level_concepts), (int(num_components),))]
            operation = pyro.sample(f"operation_{level}_{i}", dist.Categorical(torch.ones(3)))
            if operation == 0:  # Addition
                concept = lambda x: sum(c(x) for c in components)
            elif operation == 1:  # Multiplication
                concept = lambda x: torch.prod(torch.stack([c(x) for c in components]))
            else:  # Composition
                concept = lambda x: components[0](components[1](x)) if len(components) > 1 else components[0](x)
        concepts.append(concept)
    return concepts

def generate_observations(concepts: List[torch.Tensor], input_data: torch.Tensor) -> torch.Tensor:
    all_concepts = [c for level in concepts for c in level]
    num_concepts = len(all_concepts)
    weights = pyro.sample("concept_weights", dist.Dirichlet(torch.ones(num_concepts)))
    
    observations = torch.zeros_like(input_data)
    for i, concept in enumerate(all_concepts):
        observations += weights[i] * concept(input_data)
    
    noise = pyro.sample("noise", dist.Normal(0, 0.1).expand(observations.shape))
    return observations + noise