requirements.txt
# requirements.txt
pyro-ppl
torch
sympy
numpy
scipy
matplotlib